TSDK-916 Cleanup tests
